minor code review changes
